requirements.txt
pymupdf==1.24.5
Pillow==10.4.0
anthropic==0.34.2
markdown==3.7
python-frontmatter==1.1.0
numpy==2.1.0
